% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{calcCurves}
\alias{calcCurves}
\title{Calculate Relative Yield, YPR, SPR, SSB, and Recruitment curves for F/M}
\usage{
calcCurves(LB_obj)
}
\arguments{
\item{LB_obj}{An object of class \code{'LB_obj'} or class \code{'LB_pars'} that contains the life history and fishing information}
}
\value{
a dataframe with YPR, Yield, SSB, Rec, and FM
}
\description{
A function that takes a LB_pars or LB_obj object and returns a data frame of values 
for relative Yield, YPR, SPR, SSB, and Recruitment at different values of F/M
}
\author{
A. Hordyk
}

