% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDir.r
\name{DataDir}
\alias{DataDir}
\title{Report the location of the Data Files}
\usage{
DataDir()
}
\description{
A function that returns the location of the example CSV files
}
\author{
A. Hordyk modified (i.e., stolen) from T. Carruthers' code (DLMtool package)
}
