% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{getFMfun}
\alias{getFMfun}
\title{Calculate F/M given SPR and other parameters}
\usage{
getFMfun(FM, LB_pars, Control = list())
}
\arguments{
\item{FM}{a F/M value}

\item{LB_pars}{an object of class \code{'LB_pars'} that contains the life history information}

\item{Control}{a list of control options for the LBSPR model.}
}
\value{
sum of squares value
}
\description{
A internal function that optimizes for F/M when SPR is provided in the simulation parameters.
}
\details{
add Control details here
}
\author{
A. Hordyk
}

