% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineMeths.r
\docType{methods}
\name{initialize,LB_lengths-method}
\alias{initialize,LB_lengths-method}
\title{Create a new LB_lengths object}
\usage{
\S4method{initialize}{LB_lengths}(.Object, file = "none",
  LB_pars = NULL, dataType = c("raw", "freq"), header = FALSE,
  verbose = TRUE, ...)
}
\arguments{
\item{.Object}{class of object to be created}

\item{file}{file path and name to CSV containing parameters. Alternatively it can be a matrix or vector of length data}

\item{LB_pars}{a object of class LB_pars}

\item{dataType}{is the length data individual measurements (raw) or a length frequency (freq)?}

\item{header}{is there a header?}

\item{verbose}{display a message?}

\item{...}{optional additional arguments passed to read.csv}
}
\value{
a object of class \code{'LB_lengths'}
}
\description{
Function
}
\author{
A. Hordyk
}
