% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineMeths.r
\docType{methods}
\name{initialize,LB_obj-method}
\alias{initialize,LB_obj-method}
\title{Create a new LB_obj object}
\usage{
\S4method{initialize}{LB_obj}(.Object, defaults = FALSE,
  verbose = FALSE)
}
\arguments{
\item{.Object}{class of object to be created}

\item{defaults}{use defaults?}

\item{verbose}{display a message?}
}
\value{
a object of class \code{'LB_obj'}
}
\description{
Function
}
\author{
A. Hordyk
}
