% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineMeths.r
\docType{class}
\name{LB_lengths-class}
\alias{LB_lengths-class}
\title{An S4 class containing length data}
\description{
An S4 class containing length data
}
\section{Slots}{

\describe{
\item{\code{LMids}}{A numeric vector containing the mid-points of the length bins}

\item{\code{LData}}{A numeric matrix containing length data}

\item{\code{L_units}}{Character descrbing units of the length measurements}

\item{\code{Years}}{A numeric vector containing the year indices}

\item{\code{NYears}}{A length-one numeric vector for number of years}

\item{\code{Elog}}{A error log}
}}

