% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineMeths.r
\docType{class}
\name{LB_obj-class}
\alias{LB_obj-class}
\title{An S4 class containing all parameters for the LBSPR model}
\description{
An S4 class containing all parameters for the LBSPR model
}
\section{Slots}{

\describe{
\item{\code{SPR}}{The Spawning Potential Ratio}

\item{\code{Yield}}{Relative yield}

\item{\code{LMids}}{A numeric vector containing the mid-points of the length bins}

\item{\code{pLCatch}}{A numeric vector containg expected proportion for each length class in the catch}

\item{\code{pLPop}}{A numeric vector containg expected proportion for each length class in the population}

\item{\code{RelRec}}{Relative recruitment}

\item{\code{Ests}}{A matrix of estimated values}

\item{\code{NLL}}{A numeric NLL values}

\item{\code{maxFM}}{A numeric of maximum estimated F/M value (note this is apical F)}
}}

