% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LBSPR_NLLgtg}
\alias{LBSPR_NLLgtg}
\title{LBSPR Optimization function for GTG model}
\usage{
LBSPR_NLLgtg(starts, LMids, LBins, LDat, gtgLinfs, MKMat, MK, Linf, ngtg, recP,
  usePen)
}
\arguments{
\item{starts}{a}

\item{LMids}{a}

\item{LBins}{a}

\item{LDat}{a}

\item{gtgLinfs}{a}

\item{MKMat}{a}

\item{MK}{a}

\item{Linf}{a}

\item{ngtg}{a}

\item{recP}{a}

\item{usePen}{a}
}
\value{
negative log-likelihood value
}
\description{
Internal optimization function
}
\author{
A. Hordyk
}

