% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{LBSPRfit}
\alias{LBSPRfit}
\title{Fit LBSPR model to length data}
\usage{
LBSPRfit(LB_pars = NULL, LB_lengths = NULL, yrs = NA, Control = list(),
  pen = TRUE, msg = TRUE, useCPP = TRUE, ...)
}
\arguments{
\item{LB_pars}{an object of class \code{'LB_pars'} that contains the life history information}

\item{LB_lengths}{an object of class \code{'LB_lengths'} that contains the length data}

\item{yrs}{index of years to include. If NA the model is run on all years}

\item{Control}{a list of control options for the LBSPR model.}

\item{pen}{apply a penalty if estimate of selectivity is very high?}

\item{msg}{display messages?}

\item{useCPP}{use cpp optimization code?}

\item{...}{additional parameters to pass to \code{FilterSmooth}}
}
\value{
a object of class \code{'LB_obj'}
}
\description{
A function that fits the LBSPR model to length data
}
\details{
add Control details here
}
\author{
A. Hordyk
}

