% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{LBSPRfit_}
\alias{LBSPRfit_}
\title{Interal function to fit LBSPR model to length data}
\usage{
LBSPRfit_(yr = 1, LB_pars = NULL, LB_lengths = NULL, Control = list(),
  pen = TRUE, useCPP = TRUE)
}
\arguments{
\item{yr}{index of the year column to fit model to}

\item{LB_pars}{an object of class \code{'LB_pars'} that contains the life history information}

\item{LB_lengths}{an object of class \code{'LB_lengths'} that contains the length data}

\item{Control}{a list of control options for the LBSPR model.}

\item{pen}{apply a penalty if estimate of selectivity is very high?}

\item{useCPP}{use cpp optimization code?}
}
\value{
a object of class \code{'LB_obj'}
}
\description{
An internal function that fits the LBSPR model to a single year of length data
}
\details{
add Control details here
}
\author{
A. Hordyk
}

