% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{LBSPRsim}
\alias{LBSPRsim}
\title{LBSPR Simulation Model}
\usage{
LBSPRsim(LB_pars = NULL, Control = list())
}
\arguments{
\item{LB_pars}{an object of class \code{'LB_pars'} that contains the life history information}

\item{Control}{a list of control options for the LBSPR model.}
}
\value{
a object of class \code{'LB_obj'}
}
\description{
Function that generates the expected equilbrium size composition given biological parameters, and fishing mortality and selectivity pattern.
}
\details{
add Control details here
}
\author{
A. Hordyk
}

