% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{LBSPRsim_}
\alias{LBSPRsim_}
\title{Interal LBSPR Simulation Model}
\usage{
LBSPRsim_(LB_pars = NULL, Control = list(), msg = TRUE, doCheck = TRUE)
}
\arguments{
\item{LB_pars}{an object of class \code{'LB_pars'} that contains the life history information}

\item{Control}{a list of control options for the LBSPR model.}

\item{msg}{display messages?}

\item{doCheck}{check if the LB_pars object is valid? Switch off when calling function from a optimization routine.}
}
\value{
a object of class \code{'LB_obj'}
}
\description{
A internal function that generates the expected equilbrium size composition given biological parameters, and fishing mortality and selectivity pattern.  Typically only used by other functions in the package.
}
\details{
add Control details here
}
\author{
A. Hordyk
}

