% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{plotMat}
\alias{plotMat}
\title{Plot the maturity-at-length and selectivity-at-length curves}
\usage{
plotMat(LB_obj = NULL, axTex = 12, axTitle = 14, useSmooth = TRUE,
  Title = NULL)
}
\arguments{
\item{LB_obj}{an object of class \code{'LB_obj'} that contains the life history and fishing information}

\item{axTex}{size of the axis text}

\item{axTitle}{size of axis title}

\item{useSmooth}{use the smoothed estimates?}

\item{Title}{optional character string for plot title}
}
\value{
a ggplot object
}
\description{
A function that plots the maturity-at-length and selectivity-at-length curves
}
\author{
A. Hordyk
}

