% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{plotSPRCirc}
\alias{plotSPRCirc}
\title{Circle of estimated SPR and target and limit points}
\usage{
plotSPRCirc(LB_obj = NULL, SPRTarg = 0.4, SPRLim = 0.2,
  useSmooth = TRUE, Title = FALSE, Leg = TRUE)
}
\arguments{
\item{LB_obj}{an object of class \code{'LB_obj'} that contains the life history and fishing information}

\item{SPRTarg}{a numeric value specifying the SPR target}

\item{SPRLim}{a numeric value specifying the SPR limit}

\item{useSmooth}{use the smoothed estimates? Usually would want to do this}

\item{Title}{include the title?}

\item{Leg}{include the legend?}
}
\description{
A function that creates a circle plot showing the estimated SPR relative to the 
target and limit reference points
}
\author{
A. Hordyk
}

