% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{Shiny}
\alias{Shiny}
\title{Run a Shiny Application}
\usage{
Shiny(app)
}
\arguments{
\item{app}{The name of the Shiny application to run. Currently the available Shiny apps are "LBSPR" and "Sim"}
}
\description{
\code{Shiny} runs one of the Shiny Applications that are included in the package
}
\author{
A. Hordyk
}
\references{
Modified from Deal Attali's code: \url{https://www.r-bloggers.com/supplementing-your-r-package-with-a-shiny-app-2/}
}

