% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineMeths.r
\docType{methods}
\name{initialize,LB_pars-method}
\alias{initialize,LB_pars-method}
\title{Create a new LB_pars object}
\usage{
\S4method{initialize}{LB_pars}(.Object, file = "none", defaults = TRUE,
  msg = TRUE)
}
\arguments{
\item{.Object}{class of object to be created}

\item{file}{file path and name to CSV containing parameters}

\item{defaults}{use defaults for some parameters?}

\item{msg}{display a message?}
}
\value{
a object of class \code{'LB_pars'}
}
\description{
Function
}
\author{
A. Hordyk
}

