% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineMeths.r
\docType{class}
\name{LB_lengths-class}
\alias{LB_lengths-class}
\title{An S4 class containing length data}
\description{
An S4 class containing length data
}
\section{Slots}{

\describe{
\item{\code{LMids}}{Character vector of species name}

\item{\code{LData}}{A length-one numeric vector}

\item{\code{Years}}{Character vector of species name}

\item{\code{NYears}}{A length-one numeric vector}

\item{\code{Elog}}{A length-one numeric vector}
}}

