% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineMeths.r
\docType{class}
\name{LB_obj-class}
\alias{LB_obj-class}
\title{An S4 class containing all parameters for the LBSPR model}
\description{
An S4 class containing all parameters for the LBSPR model
}
\section{Slots}{

\describe{
\item{\code{SPR}}{Character vector of species name}

\item{\code{Yield}}{A length-one numeric vector}

\item{\code{LMids}}{Character vector of species name}

\item{\code{pLCatch}}{A length-one numeric vector}

\item{\code{pLPop}}{A length-one numeric vector}

\item{\code{Ests}}{A length-one numeric vector}

\item{\code{NLL}}{A length-one numeric vector}

\item{\code{maxFM}}{A length-one numeric vector}
}}

