% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineMeths.r
\docType{class}
\name{LB_pars-class}
\alias{LB_pars-class}
\title{An S4 class containing life history and other parameters}
\description{
An S4 class containing life history and other parameters
}
\section{Slots}{

\describe{
\item{\code{Species}}{Character vector of species name}

\item{\code{MK}}{A length-one numeric vector}

\item{\code{Linf}}{A length-one numeric vector}

\item{\code{CVLinf}}{A length-one numeric vector}

\item{\code{L50}}{A length-one numeric vector}

\item{\code{L95}}{A length-one numeric vector}

\item{\code{Walpha}}{A length-one numeric vector}

\item{\code{Wbeta}}{A length-one numeric vector}

\item{\code{FecB}}{A length-one numeric vector}

\item{\code{Steepness}}{A length-one numeric vector}

\item{\code{Mpow}}{A length-one numeric vector}

\item{\code{R0}}{A length-one numeric vector}

\item{\code{SL50}}{A length-one numeric vector}

\item{\code{SL95}}{A length-one numeric vector}

\item{\code{FM}}{A length-one numeric vector}

\item{\code{SPR}}{A length-one numeric vector}

\item{\code{BinMin}}{A length-one numeric vector}

\item{\code{BinMax}}{A length-one numeric vector}

\item{\code{BinWidth}}{A length-one numeric vector}
}}

