% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMat.r
\name{plotMat}
\alias{plotMat}
\title{Plot the maturity-at-length and selectivity-at-length curves}
\usage{
plotMat(
  LB_obj = NULL,
  size.axtex = 12,
  size.title = 14,
  size.leg = 12,
  useSmooth = TRUE,
  Title = NULL
)
}
\arguments{
\item{LB_obj}{an object of class \code{'LB_obj'} that contains the life history and fishing information}

\item{size.axtex}{size of the axis text}

\item{size.title}{size of axis title}

\item{size.leg}{size of legend text}

\item{useSmooth}{use the smoothed estimates?}

\item{Title}{optional character string for plot title}
}
\value{
a ggplot object
}
\description{
A function that plots the maturity-at-length and selectivity-at-length curves
}
\author{
A. Hordyk
}
