% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{plotSim}
\alias{plotSim}
\title{Plot simulated size composition}
\usage{
plotSim(LB_obj = NULL, type = c("Catch", "Pop"), Cols = NULL,
  axTex = 12, axTitle = 14)
}
\arguments{
\item{LB_obj}{an object of class \code{'LB_obj'} that contains the life history and fishing information}

\item{type}{a character value indicating if the Catch or Population should be plotted}

\item{Cols}{optional character vector of colours for the plot}

\item{axTex}{size of the axis text}

\item{axTitle}{size of axis title}
}
\value{
a ggplot object
}
\description{
A function that plots the expected size composition in the fished and unfished state
}
\author{
A. Hordyk
}

