% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSize.r
\name{plotSize}
\alias{plotSize}
\title{Plot the size data and model fits}
\usage{
plotSize(LB_obj = NULL, size.axtex = 12, size.title = 14, Title = NULL)
}
\arguments{
\item{LB_obj}{an object of class \code{'LB_obj'} that contains the life history and fishing information}

\item{size.axtex}{size of the axis text}

\item{size.title}{size of axis title}

\item{Title}{optional character string for plot title}
}
\value{
a ggplot object
}
\description{
A function that plots size data and the fitted LBSPR model
}
\author{
A. Hordyk
}
