% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSize.r
\name{plotSize}
\alias{plotSize}
\title{Plot the size data and model fits}
\usage{
plotSize(LB_obj = NULL, size.axtex = 12, size.title = 14,
  Title = NULL, scales = c("fixed", "free_x", "free_y", "free"),
  inc.text = TRUE)
}
\arguments{
\item{LB_obj}{an object of class \code{'LB_obj'} that contains the life history and fishing information}

\item{size.axtex}{size of the axis text}

\item{size.title}{size of axis title}

\item{Title}{optional character string for plot title}

\item{scales}{argument to ggplot2 function. Are scales shared across all facets
(the default, "fixed"), or do they vary across rows ("free_x"), columns ("free_y"),
or both rows and columns ("free")}

\item{inc.text}{Include text on plotting warning of high F or selectivity estimates?}
}
\value{
a ggplot object
}
\description{
A function that plots size data and the fitted LBSPR model
}
\author{
A. Hordyk
}
